Add beep-boop's JIRA monitoring to crontab
Summary:
In D15800, we added per-exercise JIRA monitoring to beep-boop.
This diff adds a cronjob to our AWS config that runs this JIRA monitoring
procedure once per hour. The 'per hour' granularity comes from the Zendesk
monitoring job which runs with that frequency; there's a chance we'll have
to tweak it if we're seeing too many false positives.

Test Plan: Verify that this is a properly formatted cronjob.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: eater

Differential Revision: https://phabricator.khanacademy.org/D15818