Rename 'internal-webserver' to 'toby'.
Summary:
I'd like to move to a scheme where the mailing list names, the
machine's hostname, and the directory-name under aws-config are all
the same.  Right now this is true for every machine but
toby/internal-webserver: the mailing list is toby-admin, but the
machine name and aws-config dir are both internal-webserver.

This change moves us to have 'toby' as the name for everything.
Consistency!

Note that the external-facing DNS names (toby.khanacademy.org) are
unchanged, and will continue to point to the same ec2 machine.  The
change in hostname just affects what 'hostname' returns when run on
the ec2 machine (and also what gets printed in the syslog/etc).

Test Plan:
Fingers crossed.  Will modify hostname manually on toby once this is
in.

Reviewers: chris

Reviewed By: chris

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D16087