Create postfix.preseed automatically.
Summary:
(Rather than requiring it to be checked in.)

I had this vision that maybe you wouldn't want mail on your machine,
and you could indicate that by not having a postfix.preseed file.  But
that's silly.  We always want mail, so we can mail cronjob failures
and the like.

This new scheme automates the mailname to be the same as the hostname,
which will be the same as the CONFIG_DIR name.  That is true
everywhere currently (it didn't used to be for
toby/internal-webserver, but is now), so the new scheme doesn't change
any existing behavior.

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D16586