Do a daily check that permissions on our passwords are ok.
A lot of people, when they're adding a new password to passphrase,
will make it world-readable (edit-policy is 'users', meaning 'all
users').  This is too permissive, as we give phabricator accounts to
contractors/etc, who shouldn't have access to secrets by default.

Now we send mail if we see something that's world-readable, so we can
manually correct it.  The right default 'wide' access is
fulltime+interns, so suggest changing it to that.

Auditors: chris

Test plan:
Ran this manually on phabricator and saw it give no output.  Took out
the grep and saw it emit a message about the 'public Wi-Fi' key.