Add a proxy for khan.co to proxy bit.ly
Summary:
This lets us have our very own url shortener!

I'm not sure exactly how to get ssl to work yet, so for right now
this is only for http.

Test Plan:
Fingers crossed.  Once this is in and the DNS change has propagated,
I'll install it on domain-redirect and see how it works!

Reviewers: desmond

Subscribers: tompryor

Differential Revision: https://phabricator.khanacademy.org/D17019