update setup scripts to require libfreetype6
Summary:
As part of the thumbnails project,
we need to dynamically composite images with the Python Imaging Library.
Drawing text with a custom font requires the libfreetype6 package.
This patch adds the dependency to the setup files.

Note that PIL will need to be recompiled/reinstalled
for the changes to take effect.
This can be triggered by running
    pip install -I pillow
at the command line.
The commit that adds PIL to requirements.txt hasn't merged yet, though,
so this shouldn't be a problem (the normal PIL update should catch it).

Test Plan:
Not sure how to test this.
The packages exist in the Ubuntu repos:
    http://packages.ubuntu.com/search?keywords=libfreetype6

Reviewers: chris, desmond

Reviewed By: desmond

Subscribers: desmond

Differential Revision: https://phabricator.khanacademy.org/D19044