Add a new ancestor_slug field to search indexes
Summary:
This makes it possible to do a query within any subtree (or set of subtrees) in
the topictree. In particular, it will be used for the "filter by domain" feature
in Android search.

Test Plan:
Ran the provisioning script on khan-academy-dev-alan .
Made some changes to the webapp code to index ancestor slugs.
Went into the cloudsearch console and ran a query with search "sine wave" and
filter query "ancestor_slugs:'math'". It gives math results on sine, and ignores
science results on waves.

Reviewers: benkomalo, johnsullivan

Reviewed By: benkomalo, johnsullivan

Subscribers: benkomalo, davidwang, hannah

Differential Revision: https://phabricator.khanacademy.org/D19955