Get nginx rotating working again.
Summary:
We occassionally get mails like this:

    /etc/cron.daily/logrotate:
    error: error running shared postrotate script for '/var/log/nginx/*.log '
    error: error running shared postrotate script for '/home/ubuntu/logs/*-access.log /home/ubuntu/logs/*-error.log '
    run-parts: /etc/cron.daily/logrotate exited with return code 1

It looks like this is a known problem with the version of nginx we use:
   https://bugs.launchpad.net/nginx/+bug/1450770

This fixes up the script based on a suggestion there.

It makes me sad to modify the system nginx logrotate file, since it
will make updating nginx more difficult.  But I don't know another way
to fix this.

Test Plan:
Fingers crossed.  Though I ran the sed command manually and it seemed
to work.

Reviewers: ethan

Reviewed By: ethan

Subscribers: davidwang

Differential Revision: https://phabricator.khanacademy.org/D19897