Change Jenkins to use HTTPS
Summary:
We're already proxying through nginx, so it's pretty easy to just change the
config.  We will also need to update Sun to make sure it's using HTTPS too; it
looks like this should be a trivial change too.  If anyone knows of other bots
or the like that might talk to Jenkins, we should update those too.

Test Plan:
Pull on jenkins, install the keys.  (Someone else may need to do this, since I
can't see K35 on Phabricator.)  Check that the web interface still works, and
that deploying via Sun still works.

Reviewers: chris

Reviewed By: chris

Subscribers: jlfwong, csilvers, alan

Differential Revision: https://phabricator.khanacademy.org/D20422