Add breadcrumbs and tutorial text as searchable fields
Summary:
Allow breadcrumbs to be searchable, and add tutorial text as a searchable field to the cloudearch domain.
Adding more data to each searchable document will hopefully allow AND searches to not over-filter things that could otherwise match the query, improving them enough to use them over OR searches.

Test Plan: Ran provisioning on khan-academy-dev-shannon (still processing). Need to link to webapp diff that fills the tutorial_text field with the description of the search item's parent tutorial or subject's description. Run test search on cloudsearch domain to make sure the tutorial_text field is filled as expected (check for articles, exercises, topics, and subjects). Next step is to add these fields to the search query and give it weights.

Reviewers: shannon, johnsullivan

Reviewed By: johnsullivan

Differential Revision: https://phabricator.khanacademy.org/D20141