Make sure that /etc isn't group-writable.
When restarting ssh, I got this warning:

  /etc/cron.daily/security-update:
  WARN: /etc is group writable!
  ssh stop/waiting
  ssh start/running, process 16031

I think it's harmless that this dir is group-writable, but I also
think there's not a reason to do it.  Let's fix that as part of the
install.

Auditors: chris

Test plan:
Fingers crossed.  I'll run this command manually on each of our
machines though.